+++ src/main/kotlin/com/wintercloud/gateway/WinterCloudGatewayApplication.kt
@@ -1,4 +1,4 @@
-package com.wintercloud.wintercloudgateway
+package com.wintercloud.gateway
 
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
